sc_custom_image/requirements.txt
scrapinghub-entrypoint-scrapy==0.12.0
selenium==3.141.0
setuptools==44.1.0
Scrapy==1.8.1
requests==2.27.1
#2captcha-python==1.1.2
pdfkit==1.0.0
boto3==1.17.36
DateTime==4.4
Pillow==6.2.2
TwoCaptcha==0.0.1
#tika
numpy==1.16.6
PyPDF2==1.27.8
captcha_solver==0.1.5
#chromedriver-py==98.0.4758.102
webdriver_manager